Change the printer so it no longer holds the Banner as state
